Fix: chating 모달 구조 수정
