+++ beginner-gain-client/src/pages/make-boilerplate/etc.tsx
@@ -1,4 +1,5 @@
 import Screen from '@/components/screen/make-bolierplate/etc';
+import Chat from "@/components/screen/chat";
 
 export const getServerSideProps = async ({ query = {}, params = {} }) => {
   return {
@@ -9,4 +10,14 @@ export const getServerSideProps = async ({ query = {}, params = {} }) => {
   };
 };
 
-export default Screen;
+const Page = () => {
+
+  return (
+      <>
+        <Chat/>
+        <Screen />
+      </>
+  );
+};
+
+export default Page;
